Factored out LFeatures into LFeaturesSets
